requirements.txt
# Google Generative AI
google-generativeai==0.8.3
google-cloud-aiplatform==1.73.0

# Core LangChain (for vector DB only)
langchain==0.3.13
langchain-community==0.3.13
langchain-chroma==0.1.4

# PDF processing
pypdf==5.1.0

# Vector database
chromadb==0.5.23

# Alternative LLM support
openai==1.59.6
langchain-openai==0.2.14

# Environment variables
python-dotenv==1.0.1

# Utilities
tiktoken==0.8.0

# Observability & Logging
structlog==24.1.0
colorlog==6.8.2

# Performance & Caching
cachetools==5.3.2

# Data validation - UPDATED VERSION
pydantic>=2.7.4,<3.0.0

# Rich console output
rich==13.7.0

# HTTP requests
requests==2.31.0